requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.3.1
pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.4.2
jinja2>=3.1.2
aiofiles>=23.2.1
pytest>=7.4.3
httpx>=0.25.1
pdf2image>=1.16.3
poppler-utils>=23.01.0